PE-142: Integrate Karaf and CM: clean up
