+++ features/config/upgrade/src/test/java/liquibase/ext2/cm/change/types/NumberTypeTest.java
@@ -29,6 +29,7 @@
 package liquibase.ext2.cm.change.types;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 import static org.opennms.core.test.OnmsAssert.assertThrowsException;
 
 import java.util.List;
@@ -86,11 +87,11 @@ public class NumberTypeTest extends AbstractTypeTest {
     }
 
     @Test
-    public void shouldDefaultTo0ForNoDefaultValue() {
+    public void shouldDefaultToNullForNoDefaultValue() {
         attributes.remove("default");
         ConfigItem item = createItem();
         assertEquals(Type.NUMBER, item.getType());
-        assertEquals((double) 0, item.getDefaultValue());
+        assertNull(item.getDefaultValue());
     }
 
     @Override
