+++ features/datachoices/src/main/java/org/opennms/features/datachoices/internal/StateManager.java
@@ -36,7 +36,7 @@ import java.util.UUID;
 
 import org.opennms.features.config.service.api.ConfigUpdateInfo;
 import org.opennms.features.config.service.api.ConfigurationManagerService;
-import org.opennms.features.config.service.util.PropertiesCacheWithCm;
+import org.opennms.features.config.service.util.CmProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -65,44 +65,43 @@ public class StateManager {
     private static final String ACKNOWLEDGED_AT_KEY = "acknowledged-at";
 
     private final List<StateChangeHandler> m_listeners = Lists.newArrayList();
-    private final PropertiesCacheWithCm m_propertiesCache;
-    private final ConfigUpdateInfo configKey;
+    private final CmProperties propertiesCache;
 
     public interface StateChangeHandler {
         void onIsEnabledChanged(boolean isEnabled);
     }
 
     public StateManager(ConfigurationManagerService cm) {
-        configKey = new ConfigUpdateInfo(PROPERTIES_CONFIG_NAME, "default");
-        m_propertiesCache = new PropertiesCacheWithCm(cm);
+        ConfigUpdateInfo configIdentifier = new ConfigUpdateInfo(PROPERTIES_CONFIG_NAME, "default");
+        propertiesCache = new CmProperties(cm, configIdentifier);
     }
 
     public Boolean isEnabled() throws IOException {
-        return (Boolean) m_propertiesCache.getProperty(configKey, ENABLED_KEY);
+        return (Boolean) propertiesCache.getProperty(ENABLED_KEY);
     }
 
     public void setEnabled(boolean enabled, String user) throws Exception {
-        m_propertiesCache.setProperty(configKey, ENABLED_KEY, enabled);
-        m_propertiesCache.setProperty(configKey, ACKNOWLEDGED_BY_KEY, user == null ? "" : user);
-        m_propertiesCache.setProperty(configKey, ACKNOWLEDGED_AT_KEY, new Date().toString());
+        propertiesCache.setProperty(ENABLED_KEY, enabled);
+        propertiesCache.setProperty(ACKNOWLEDGED_BY_KEY, user == null ? "" : user);
+        propertiesCache.setProperty(ACKNOWLEDGED_AT_KEY, new Date().toString());
         for (StateChangeHandler listener : m_listeners) {
             listener.onIsEnabledChanged(enabled);
         }
     }
 
     public String getOrGenerateSystemId() throws IOException {
-        String systemId = (String)m_propertiesCache.getProperty(configKey, SYSTEM_ID_KEY);
+        String systemId = (String) propertiesCache.getProperty(SYSTEM_ID_KEY);
         if (systemId == null) {
             LOG.debug("No existing system id was found. Generating a new system id.");
             systemId = UUID.randomUUID().toString();
-            m_propertiesCache.setProperty(configKey, SYSTEM_ID_KEY, systemId);
+            propertiesCache.setProperty(SYSTEM_ID_KEY, systemId);
         }
         return systemId;
     }
 
     public String getAndRegenerateSystemId() throws IOException {
         String systemId = UUID.randomUUID().toString();
-        m_propertiesCache.setProperty(configKey, SYSTEM_ID_KEY, systemId);
+        propertiesCache.setProperty(SYSTEM_ID_KEY, systemId);
         return systemId;
     }
